Convert frame angles to degrees
This normalises TR1-3 and TR4/5 frame rotations by handling everything in degrees.

+++ TRLevelControl/Model/Common/TRAnimFrameRotation.cs
@@ -3,9 +3,9 @@
 public class TRAnimFrameRotation : ICloneable
 {
     public TRAngleMode Mode { get; set; } = TRAngleMode.Auto;
-    public short X { get; set; }
-    public short Y { get; set; }
-    public short Z { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
+    public float Z { get; set; }
 
     public TRAnimFrameRotation Clone()
         => (TRAnimFrameRotation)MemberwiseClone();
